refactor: improve employees page structure and components, update currency to PHP in dashboard
