+++ src/pages/Dashboard.jsx
@@ -107,9 +107,9 @@ const Dashboard = () => {
   const PIE_COLORS = ["#f59e0b", "#06b6d4", "#10b981", "#ef4444", "#a855f7", "#ec4899"]
 
   const formatCurrency = (value) => {
-    return new Intl.NumberFormat("en-US", {
+    return new Intl.NumberFormat("en-PH", {
       style: "currency",
-      currency: "USD",
+      currency: "PHP",
       minimumFractionDigits: 0,
     }).format(value)
   }
